Rename com.paypal.android.ui.card package to com.paypal.android.ui.approveorder.
