+++ Demo/src/main/java/com/paypal/android/ui/approveorder/ApproveOrderUiState.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.ui.card
+package com.paypal.android.ui.approveorder
 
 import androidx.compose.runtime.Immutable
 import com.paypal.android.api.model.Order
