+++ Demo/src/main/java/com/paypal/android/ui/approveorder/ApproveOrderView.kt
@@ -1,4 +1,4 @@
-package com.paypal.android.ui.card
+package com.paypal.android.ui.approveorder
 
 import android.content.Context
 import android.content.ContextWrapper
