+++ Demo/src/main/res/navigation/nav_graph.xml
@@ -26,7 +26,7 @@
     </fragment>
     <fragment
         android:id="@+id/cardFragment"
-        android:name="com.paypal.android.ui.card.ApproveOrderFragment"
+        android:name="com.paypal.android.ui.approveorder.ApproveOrderFragment"
         android:label="CardFragment">
         <action
             android:id="@+id/action_cardFragment_to_selectCardFragment"
